use int64_t instead of long long

+++ mmdet3d/ops/iou3d/src/iou3d.cpp
@@ -12,6 +12,7 @@ All Rights Reserved 2019-2020.
 #include <torch/extension.h>
 #include <torch/serialize/tensor.h>
 
+#include <cstdint>
 #include <vector>
 
 #define CHECK_CUDA(x) \
@@ -103,7 +104,7 @@ int nms_gpu(at::Tensor boxes, at::Tensor keep,
 
   int boxes_num = boxes.size(0);
   const float *boxes_data = boxes.data_ptr<float>();
-  long long *keep_data = keep.data_ptr<long long>();
+  int64_t *keep_data = keep.data_ptr<int64_t>();
 
   const int col_blocks = DIVUP(boxes_num, THREADS_PER_BLOCK_NMS);
 
@@ -124,7 +125,7 @@ int nms_gpu(at::Tensor boxes, at::Tensor keep,
 
   cudaFree(mask_data);
 
-  unsigned long long* remv_cpu = new unsigned long long[col_blocks]();
+  unsigned long long *remv_cpu = new unsigned long long[col_blocks]();
 
   int num_to_keep = 0;
 
@@ -157,7 +158,7 @@ int nms_normal_gpu(at::Tensor boxes, at::Tensor keep,
 
   int boxes_num = boxes.size(0);
   const float *boxes_data = boxes.data_ptr<float>();
-  long long *keep_data = keep.data_ptr<long long>();
+  int64_t *keep_data = keep.data_ptr<int64_t>();
 
   const int col_blocks = DIVUP(boxes_num, THREADS_PER_BLOCK_NMS);
 
@@ -178,7 +179,7 @@ int nms_normal_gpu(at::Tensor boxes, at::Tensor keep,
 
   cudaFree(mask_data);
 
-  unsigned long long* remv_cpu = new unsigned long long[col_blocks]();
+  unsigned long long *remv_cpu = new unsigned long long[col_blocks]();
 
   int num_to_keep = 0;
